+++ docs/pages/docs/guide/advanced/playground.mdx
@@ -27,7 +27,7 @@ Due to the purely client-side nature of this component, there are some features 
 - [x] Frontmatter
 - [x] Mermaid.js
 
-For more information, see \`compileMdxClient.ts\`.`)
+For more information, see \`compile-mdx-client.ts\`.`)
   return (
     <div className="grid grid-cols-2 gap-2 mt-6">
       <textarea
